LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/popover-implicit-anchor.tentative-expected.txt

Button Popover

FAIL Popover invokers form an implicit anchor reference assert_equals: expected 100 but got 55
